refactor: code relocations
