+++ src/components/RegistrationForm/Inputs/index.tsx
@@ -1,9 +1,9 @@
-import { FormFieldOption } from '../form-fields';
+import { FormFieldOptions } from '../form-field';
 import Input from './Input';
 import InputsList from './InputsList';
 import Select from './Select';
 
-const renderInput = (options: FormFieldOption) => {
+const renderInput = (options: FormFieldOptions) => {
   const { type, inputRef, formFieldType } = options;
   const ref = (inputRef !== undefined && !Array.isArray(inputRef) && inputRef.ref) || null;
   switch (type) {
