+++ src/components/RegistrationForm/form-field/index.ts
@@ -1,0 +1,7 @@
+import FormField from './FormField';
+import { InputType, FormFieldType, FormFieldBaseOptions, FormFieldOptions } from './types';
+import { FormFieldOptionsContext, withFormFieldOptions } from './context';
+import { FieldName, formFields } from './config';
+
+export type { InputType, FormFieldType, FormFieldBaseOptions, FormFieldOptions, FieldName };
+export { FormField, FormFieldOptionsContext, formFields, withFormFieldOptions };
